przeniesienie klasy Node do  klasy StackList
